Improves error-handling for GitHub requests
We're seeing intermittent 500 errors from the cache-npm-wrangler-binary
Worker. Logs tell us that a call to `getJSONFromGitHub` is catching and
re-throwing a "Fetch API cannot load: undefined" error.

This could conceivably occur if `getReleaseByTag(tag)` returns a
response that does not have an `assets_url` property.

This changes logs the response from GitHub if the response status is not
200 or if `release` does not have an `assets_url` property.